refactor(android): extract Snapshot3D enter/exit JS into bundled assets
Mirrors the macOS Snapshot3DBridge split: the long 3D-inspector script
bodies move from Kotlin raw-string constants into JS files under
`app/src/main/assets/snapshot3d/`. The enter script lives as four phase
files (setup, collect, apply, input), concatenated at first access so
each individual source file stays within the project's 500-line limit.
The phases share closure state inside one IIFE, so the concatenation is
what runs.

before: Snapshot3DBridge.kt 749 lines
after:
  Snapshot3DBridge.kt                   67
  assets/snapshot3d/enter-setup.js     101
  assets/snapshot3d/enter-collect.js   133
  assets/snapshot3d/enter-apply.js      93
  assets/snapshot3d/enter-input.js     333
  assets/snapshot3d/exit.js             57
